requirements.txt

streamlit
openai
python-docx
pdfplumber
